+++ scripts/automation/Radius/Changelog.md
@@ -14,6 +14,7 @@ This release offers a significant overhaul for the Radius Cert Deployment tool, 
 - Association data is cached up front rather than gathered throughout the script, offering performance improvements for organizations with a large number of radius users
 - Added ability to run each public function headless in order to automate cert generation and distribution
 - Added a table to keep track of generated/deployed certificates when using the tool
+- Added password validation (re-enter) when generating root certificate
 
 ## 1.1.0